backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
supabase==1.0.4
httpx==0.25.0
uuid==1.30
pyjwt==2.8.0
pytest==7.4.3 